demo/requirements.txt
faker~=23.1.0
polars~=0.20.0
pyarrow~=15.0.0
neo4j~=5.18.0
python-dotenv==1.0.1
codetiming>=1.4.0
